Move template string parsing to quotes
Refs #421
